move VLM shim tests to provider packages
